# src/main/resources/static/swagger.yml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/sudhirtumati/countries/1.0.0
info:
  description: Provides information about <a href="http://www.nationsonline.org/oneworld/country_code_list.htm">Countries</a>
  version: "1.0.0"
  title: Countries API
  contact:
    email: mailofsudhir@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
tags:
  - name: countries
paths:
  /search:
    get:
      tags:
        - countries
      summary: Search for a country
      operationId: search
      description: Full text search for a country
      parameters:
        - in: query
          name: q
          description: Search criteria
          required: true
          schema:
            type: string
      responses:
        '200':
          description: countries matching with provided search criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Country'
        '400':
          description: no search criteria provided
  /countries/{countryId}:
    get:
      tags:
        - countries
      summary: Get details of a country
      operationId: get
      description: |
        Returns details of a country identified with provided identifier
      parameters:
        - in: path
          name: countryId
          description: ISO3 code of the country
          required: true
          schema:
            type: string
            minLength: 3
            maxLength: 3
      responses:
        '200':
          description: Details of a country identified with provided identifier
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/Country'
        '404':
          description: No country found with provided identifier
  /countries:
    get:
      tags:
        - countries
      summary: Get list of countries
      operationId: get
      description: |
        Returns list of all countries if no filters are provided. Otherwise returns list of countries matching with provided filters.
      parameters:
        - in: query
          name: countryId
          description: ISO3 code of the country
          required: false
          schema:
            type: string
            minLength: 3
            maxLength: 3
        - in: query
          name: iso2
          description: ISO2 code of the country
          schema:
            type: string
            minLength: 2
            maxLength: 2
        - in: query
          name: numericCode
          description: ISO numeric code of the country
          schema:
            type: string
            minLength: 3
            maxLength: 3
      responses:
        '200':
          description: all countries or countries matching with provided criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Country'
        '400':
          description: two or more input parameters provided
    post:
      tags:
        - countries
      summary: Add a new country
      operationId: add
      description: Add a new country to the list
      responses:
        '201':
          description: country created
        '400':
          description: 'invalid input, object invalid'
        '409':
          description: Country already exists
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Country'
        description: Country to add

components:
  schemas:
    Country:
      type: object
      required:
        - countryId
        - name
        - iso2
        - numericCode
      properties:
        countryId:
          type: string
          minLength: 3
          maxLength: 3
          example: IND
        name:
          type: string
          example: India
        iso2:
          type: string
          minLength: 2
          maxLength: 2
          example: IN
        numericCode:
          type: string
          minLength: 3
          maxLength: 3
          example: 356